feat(keys): 现在可以自定义返回的自增主键类型。
现在可以通过 returnGeneratedKey(Class<T> numberClass) 方法要求返回指定类型的自增主键。

BREAKING CHANGE: 移除了对于“是否返回主键”的选择，一旦定义了主键类型，就代表action将返回该类型的主键。

+++ api/src/main/java/cc/carm/lib/easysql/api/builder/DeleteBuilder.java
@@ -2,7 +2,7 @@ package cc.carm.lib.easysql.api.builder;
 
 import cc.carm.lib.easysql.api.SQLAction;
 
-public interface DeleteBuilder extends ConditionalBuilder<DeleteBuilder, SQLAction<Long>> {
+public interface DeleteBuilder extends ConditionalBuilder<DeleteBuilder, SQLAction<Integer>> {
 
     String getTableName();
